lift and shift Buffer into a separate module
