+++ upstairs/src/test.rs
@@ -821,25 +821,4 @@ pub(crate) mod up_test {
         // Back to being below the limit
         assert!(wr.send_io_live_repair(Some(3)));
     }
-
-    #[test]
-    fn block_eat_without_memcpy() {
-        let mut upper = Buffer::new(10, 512);
-        let mut lower = Buffer::new(10, 512);
-
-        // Write block 1
-        lower.write(512, &[0xFE; 512]);
-        let prev_ptr = upper.as_ptr();
-
-        // Eating a partially owned buffer requires a memcpy
-        let _ = upper.eat(0, lower);
-        assert_eq!(prev_ptr, upper.as_ptr());
-
-        // Eating the entire owned buffer just swaps pointers
-        let mut lower = Buffer::new(10, 512);
-        lower.write(0, &[0xFE; 512 * 10]);
-        let prev_ptr = lower.as_ptr();
-        let _ = upper.eat(0, lower);
-        assert_eq!(prev_ptr, upper.as_ptr());
-    }
 }
